feat: move error code logic
