+++ libs/ledgerjs/packages/hw-app-exchange/src/ReturnCode.ts
@@ -30,7 +30,69 @@ export const ErrorStatus = {
   INTERNAL_ERROR: 0x6a85,
   WRONG_P1: 0x6a86,
   WRONG_P2: 0x6a87,
-  CLASS_NOT_SUPPORTED: 0x6e00,
+  WRONG_P2_EXTENSION: 0x6a88,
+  INVALID_P2_EXTENSION: 0x6a89,
   INVALID_INSTRUCTION: 0x6d00,
+  UNEXPECTED_INSTRUCTION: 0x6d01,
+  CLASS_NOT_SUPPORTED: 0x6e00,
+  MALFORMED_APDU: 0x6e01,
+  INVALID_DATA_LENGTH: 0x6e02,
   SIGN_VERIFICATION_FAIL: 0x9d1a,
 } as const;
+
+// const ErrorMessage: Record<keyof typeof ErrorStatus, string> = {
+//   INCORRECT_COMMAND_DATA: "Incorrect command data",
+//   DESERIALIZATION_FAILED: "Payload deserialzation failed",
+//   WRONG_TRANSACTION_ID: "Wrond transaction id",
+//   INVALID_ADDRESS: "Invalid address",
+//   USER_REFUSED: "User refused",
+//   INTERNAL_ERROR: "Internal error",
+//   WRONG_P1: "Wrong P1",
+//   WRONG_P2: "Wrong P2",
+//   WRONG_P2_EXTENSION: "",
+//   INVALID_P2_EXTENSION: "",
+//   INVALID_INSTRUCTION: "Invalid instruction",
+//   UNEXPECTED_INSTRUCTION: "",
+//   CLASS_NOT_SUPPORTED: "Class not supported",
+//   MALFORMED_APDU: "",
+//   INVALID_DATA_LENGTH: "",
+//   SIGN_VERIFICATION_FAIL: "Signature verification failed",
+// };
+
+export function getExchageErrorMessage(errorCode: number): string | undefined {
+  switch (errorCode) {
+    case ErrorStatus.INCORRECT_COMMAND_DATA:
+      return "Incorrect command data";
+    case ErrorStatus.DESERIALIZATION_FAILED:
+      return "Payload deserialzation failed";
+    case ErrorStatus.WRONG_TRANSACTION_ID:
+      return "Wrond transaction id";
+    case ErrorStatus.INVALID_ADDRESS:
+      return "Invalid address";
+    case ErrorStatus.USER_REFUSED:
+      return "User refused";
+    case ErrorStatus.INTERNAL_ERROR:
+      return "Internal error";
+    case ErrorStatus.WRONG_P1:
+      return "Wrong P1";
+    case ErrorStatus.WRONG_P2:
+      return "Wrong P2";
+    case ErrorStatus.WRONG_P2_EXTENSION:
+      return "";
+    case ErrorStatus.INVALID_P2_EXTENSION:
+      return "";
+    case ErrorStatus.UNEXPECTED_INSTRUCTION:
+      return "";
+    case ErrorStatus.CLASS_NOT_SUPPORTED:
+      return "Class not supported";
+    case ErrorStatus.INVALID_INSTRUCTION:
+      return "Invalid instruction";
+    case ErrorStatus.MALFORMED_APDU:
+      return "";
+    case ErrorStatus.INVALID_DATA_LENGTH:
+      return "";
+    case ErrorStatus.SIGN_VERIFICATION_FAIL:
+      return "Signature verification failed";
+  }
+  return undefined;
+}
